- Deleted Adventure Theme  - very similar to Adapt, just different background  - moved the background into apt/images/folder
- Post Drafts!
 - auto-saves new posts while you are writing them
 - auto-saves edits of existing posts
  - 'edit draft' and 'edit post' will appear for these thanks to new $post['edit']
  - all themes except doc updated tly handle draft editing
 - so much more, this will require further testing, but I think it's done for the most part

- Code clean-up
 - fixed many one line if statements that did not have {and  }

- Changes
 - Only 1 post can time now (this was probably the case before, but EditSee prevents it now)
 - draft field is altered all the time currently to b ther  tinyint
 = stuff I forgot...

-Apex

+++ trunk/includes/database/editsee_Database.create.php
@@ -77,7 +77,7 @@ $project7->db->_query("CREATE TABLE IF NOT EXISTS `".$table_prefix."post` (
   `type` char(4) NOT NULL,
   `in_nav` tinyint(1) NOT NULL,
   `page_order` int(11) NOT NULL,
-  `draft` tinyint(1) NOT NULL,
+  `draft` int(11) NOT NULL,
   `date_entered` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
   `date_deleted` timestamp NOT NULL DEFAULT '0000-00-00 00:00:00',
   `deleted` tinyint(1) NOT NULL,
